cplusco sudirman v.2.1

+++ application/models/files_model.php
@@ -29,8 +29,10 @@ class Files_model extends CI_Model
 		return $data;
 	}
 
-	public function get_file($file_id = 0)
+	public function get_file($file_id)
 	{
+		return $this->data_files('*', array('id_files' => $file_id))->row_array();
+
 		/*$query = 'SELECT * FROM master_files ';
 		if($file_id > 0)
 		{
@@ -43,12 +45,21 @@ class Files_model extends CI_Model
 	}
 	public function remove_files($where)
 	{
-		$this->db->delete('master_files', $where); 
+		$file = $this->data_files('*', $where);
+		$files = $file->result_array();
+		if(count($files > 0 ) )
+		{
+			foreach ($files as $key => $value) {
+				$loc = $value['file_path'].''.$value['file_name'];
+				unlink( $loc );
+			}
+			$this->db->delete('master_files', $where); 
+		}
 	}
 
 	public function save_file($data)
 	{
-		$this->db->query('INSERT INTO master_files(file_name, file_type, file_path, raw_name, original_name, client_name, file_ext, file_size) values(?,?,?,?,?,?,?,?)',  array($data['file_name'], $data['file_type'], $data['file_path'], $data['raw_name'], $data['orig_name'], $data['client_name'], $data['file_ext'], $data['file_size']));
+		$this->db->query('INSERT INTO master_files(file_name, file_type, file_path, raw_name, original_name, client_name, file_ext, file_size, file_url) values(?,?,?,?,?,?,?,?,?)',  array($data['file_name'], $data['file_type'], $data['file_path'], $data['raw_name'], $data['orig_name'], $data['client_name'], $data['file_ext'], $data['file_size'], $data['file_url']));
 		return $this->db->insert_id();
 	}
 
@@ -70,6 +81,12 @@ class Files_model extends CI_Model
 				$file = $this->upload->data();
 				$file['file_url'] = base_url($config['upload_path'].$file['file_name']);
 				$insert_id = $this->save_file($file);
+				
+
+				$download_url = base_url('files/download_file/'.$insert_id);
+				$this->db->where(array('id_files' => $insert_id));
+				$this->db->update('master_files', array('download_url' => $download_url));
+
 				$file['id_files'] = $insert_id;
 				$key = explode('-', $key);
 				if(count($key) > 1)
